Change text for days since installation (EXPOSUREAPP-6574) (#2969)
* add strings for 0 and 1 day since installation

* klint

* change calculation

* fix test

* use get

+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/tracing/ui/details/items/periodlogged/PeriodLoggedBox.kt
@@ -33,12 +33,12 @@ class PeriodLoggedBox(
     }
 
     data class Item(
-        val daysSinceInstallation: Long,
+        val daysSinceInstallation: Int,
         val tracingStatus: GeneralTracingStatus.Status
     ) : DetailsItem {
 
         fun getInstallTimePeriodLogged(context: Context): String =
-            if (daysSinceInstallation < 14L) {
+            if (daysSinceInstallation < 14) {
                 context.getString(
                     R.string.risk_details_information_body_period_logged_assessment_under_14_days
                 ).format(daysSinceInstallation)
